+++ library/src/commonTest/kotlin/io/grafima/charts/ColorContrastTest.kt
@@ -20,6 +20,8 @@ import androidx.compose.ui.graphics.Color
 import io.grafima.charts.bar.AxisConfig
 import io.grafima.charts.bar.ChartStyle
 import io.grafima.charts.bar.TooltipSelectionRenderer
+import io.grafima.charts.line.LegendLabelTextStyle
+import io.grafima.charts.line.LineAxisConfig
 import io.grafima.charts.line.LineCrosshairConfig
 import io.grafima.charts.pie.TooltipPieSelectionRenderer
 import io.grafima.charts.radar.TooltipRadarSelectionRenderer
@@ -103,6 +105,8 @@ class ColorContrastTest {
         assertMeetsAA("Bar x-axis label", style.labelTextStyle.color, Color.White)
         assertMeetsAA("Bar value label", style.valueTextStyle.color, Color.White)
         assertMeetsAA("Y-axis label", AxisConfig().axisLabelTextStyle.color, Color.White)
+        assertMeetsAA("Line axis label", LineAxisConfig().labelColor, Color.White)
+        assertMeetsAA("Line legend label", LegendLabelTextStyle.color, Color.White)
     }
 
     private companion object {